Redesign ActiveProjectService
Due to problems with the desynchronization
of the stopwatch UI and data, the logic had
to be completely redesigned.

However, the problem could not be solved in a completely normal way.

The error occurs when summing the elapsed time.
The more summation, the greater the error.
As a result, a difference of a couple of milliseconds
makes a noticeable difference between the UI and the data.

To solve the problem, I had to get the elapsed time
from the database and transfer it to the UI.

Not the best solution, but I don't see any other solutions.