AboutPageStyles.licenses - Add margin bottom
Not sure if this is the right decision.

It's strange that the button panel in Android
overlaps the application interface.

It turns out that if the panel is high,
it will still block the application interface.
Or, on the contrary, the indentation
will be too large if the panel does not overlap the application.